sample: 适配各 sample 到 BM1684X2(SE13-64)
补建/补测 sail C++ 例程(segformer/YOLOv5/GroundingDINO/BERT)并修正 README 脚注:
- segformer: 加 SE13-64 segformer_sail.soc 精度行(fp16/int8), 脚注改写
- YOLOv5: 加 SE13-64 bmcv.soc/sail.soc 精度+性能行, 脚注由"硬件解码崩溃"改为"libsophon 0.4.13 已修复"
- GroundingDINO: 加 SE13-64 groundingdino_sail.soc 性能行
- BERT: 加 SE13-64 bert_sail.soc 精度+性能行, 脚注改写
- ResNet: vpp 修复后重测硬件解码路径, bmcv.py/opencv.soc/bmcv.soc 精度由~69%恢复至~80%(与软件解码一致), int8_4b err=9 已修复; README 精度行+性能行(bmcv.py int8_4b)+脚注+auto_test 预期值同步更新
- auto_test.sh: 各 sample BM1684X2 段补 test_cpp/eval_cpp sail/bmcv 用例

注: Qwen3_5 适配已在别处独立完成, 本提交不含 Qwen3_5 改动。

(cherry picked from commit 0fa577b3882c0e5e42f8caa069e9e84598bed0a5)

+++ sample/HRNet_pose/python/hrnet_pose.py
@@ -84,7 +84,8 @@ def save_coco_keypoints(keypoints, maxvals, image_name, box_score):
     res = {"image_id": image_id,
            "category_id": 1,  # person
            "keypoints": keypoints,
-           "score": box_score * k_score}
+           # np.float32 无法被标准库 json 序列化，此处显式转为 Python float
+           "score": float(box_score * k_score)}
     return res
